Project is now a library with main-like function to enable doc testing
